# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Eligo Backend API
  version: 1.7.0
  description: |
    Eligo 后端接口的唯一机器可读契约。
    M1 个人资料契约采用“同一账号叠加能力”模型：普通参与和个人发布不拆分账号身份，
    用户与后端预置企业建立有效 OWNER 关系后，再叠加企业操作能力。
    当前不提供企业申请、认证、审核或身份切换接口。
    新增或修改接口时，后端开发者必须在同一个提交中更新本文件。
    除文件内容下载接口外，所有 JSON 响应都使用统一 Result 包装。
  x-generated-from-code-date: 2026-08-22
  x-contract-updated-date: 2026-08-22
  x-m1-contract-status: REVIEW
  x-m2-contract-status: IMPLEMENTED
  x-m3-contract-status: IMPLEMENTED
  x-m4-contract-status: IMPLEMENTED
  x-m4.1-contract-status: IMPLEMENTED
servers:
  - url: http://127.0.0.1:8080
    description: 本机 Docker / 微信开发者工具
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/activities:
    get:
      tags: [Activity]
      operationId: listPublicActivities
      summary: 游标分页查询公开活动
      description: |
        匿名可访问。默认只返回 PUBLISHED；显式状态筛选还可查询 CANCELLED 或 ENDED。
        活动到达 endsAt 后立即按 ENDED 返回，并从 PUBLISHED 列表移入 ENDED 列表。
        keyword 对标题或描述执行包含匹配；topic 对规范化话题名称执行精确匹配，并与状态、分类、区县和关键词筛选按“且”共同生效。
        未提供 userLatitude 与 userLongitude 时固定按 startsAt ASC、activityId ASC 排序。
        两个用户坐标必须同时提供并使用 GCJ-02；提供后以 Haversine 公式计算整数米距离，按
        distanceMeters ASC、activityId ASC 排序。radiusMeters 仅允许在用户坐标完整时提供。
        距离模式 cursor 与用户坐标、radiusMeters、状态、分类、区县、关键词等筛选条件绑定；
        改变筛选条件后复用旧游标返回 10001/400。cursor 是不透明字符串，客户端不得解析或改写。
      x-implementation-status: IMPLEMENTED
      security: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/ActivityLimit"
        - $ref: "#/components/parameters/PublicActivityStatus"
        - name: categoryCode
          in: query
          required: false
          description: 活动单选主分类精确匹配。
          schema:
            $ref: "#/components/schemas/ActivityCategory"
        - name: regionCode
          in: query
          required: false
          description: 区县级行政区划编码精确匹配。
          schema:
            $ref: "#/components/schemas/ActivityRegionCode"
        - name: keyword
          in: query
          required: false
          description: 标题或描述包含匹配；去除首尾空白后最多 20 个 Unicode 码点，空白等同未传。
          schema:
            type: string
            maxLength: 20
        - name: topic
          in: query
          required: false
          description: 活动自由文本话题精确匹配；去除首尾空白并忽略大小写，最多 20 个 Unicode 码点。
          schema:
            type: string
            minLength: 1
            maxLength: 20
        - name: userLatitude
          in: query
          required: false
          description: 用户当前位置纬度；必须与 userLongitude 同时提供，使用 GCJ-02，服务端不保存。
          schema:
            $ref: "#/components/schemas/ActivityLatitude"
        - name: userLongitude
          in: query
          required: false
          description: 用户当前位置经度；必须与 userLatitude 同时提供，使用 GCJ-02，服务端不保存。
          schema:
            $ref: "#/components/schemas/ActivityLongitude"
        - name: radiusMeters
          in: query
          required: false
          description: 以用户坐标为圆心的精确距离过滤半径；未提供完整用户坐标时不得传入。
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 50000
      responses:
        "200":
          description: 公开活动分页
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicActivityPageResult"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/activities/map:
    get:
      tags: [Activity]
      operationId: listActivitiesOnMap
      summary: 查询地图视野内的公开活动
      description: |
        匿名可访问。只返回数据库状态为 PUBLISHED、尚未到达 endsAt 且坐标完整的活动。
        已取消、已结束、已隐藏或无坐标活动不返回。接口支持两个互斥模式：
        视窗模式必须同时提供四个边界；可选完整 userLatitude/userLongitude，提供后按 distanceMeters ASC、activityId ASC 排序，
        否则按 startsAt ASC、activityId ASC 排序。附近模式必须省略四个边界，同时提供用户坐标和 radiusMeters，
        按 Haversine 整数米距离过滤并按 distanceMeters ASC、activityId ASC 排序。部分边界、单个用户坐标、
        四个边界与 radiusMeters 混用均返回 10001/400。
        响应最多返回 limit 条；truncated=true 表示当前视野内仍有更多结果，客户端应缩小地图范围后重新查询。
        请求边界与响应坐标统一使用 GCJ-02（国测局坐标系），不得直接混入 WGS84 或 BD-09 坐标。
        本接口不保存用户位置，不返回预计车程、天气或路线规划。
      x-implementation-status: IMPLEMENTED
      security: []
      parameters:
        - name: minLatitude
          in: query
          required: false
          description: 地图视野最小纬度；视窗模式下四个边界必须同时提供，附近模式下必须省略。
          schema:
            $ref: "#/components/schemas/ActivityLatitude"
        - name: maxLatitude
          in: query
          required: false
          description: 地图视野最大纬度，必须大于 minLatitude。
          schema:
            $ref: "#/components/schemas/ActivityLatitude"
        - name: minLongitude
          in: query
          required: false
          description: 地图视野最小经度；视窗模式下四个边界必须同时提供，附近模式下必须省略。
          schema:
            $ref: "#/components/schemas/ActivityLongitude"
        - name: maxLongitude
          in: query
          required: false
          description: 地图视野最大经度，必须大于 minLongitude；首版不支持跨越国际日期变更线的视野。
          schema:
            $ref: "#/components/schemas/ActivityLongitude"
        - name: userLatitude
          in: query
          required: false
          description: 用户当前位置纬度；必须与 userLongitude 同时提供，统一使用 GCJ-02，服务端不保存。
          schema:
            $ref: "#/components/schemas/ActivityLatitude"
        - name: userLongitude
          in: query
          required: false
          description: 用户当前位置经度；必须与 userLatitude 同时提供，统一使用 GCJ-02，服务端不保存。
          schema:
            $ref: "#/components/schemas/ActivityLongitude"
        - name: radiusMeters
          in: query
          required: false
          description: 附近模式必填的精确距离半径；不得与四个视窗边界混用。
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 50000
        - name: categoryCode
          in: query
          required: false
          description: 活动单选主分类精确匹配。
          schema:
            $ref: "#/components/schemas/ActivityCategory"
        - name: limit
          in: query
          required: false
          description: 单个地图视野最多返回的活动数量。
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 200
            default: 100
      responses:
        "200":
          description: 地图视野内的公开活动
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityMapResult"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/activities/{activityId}:
    parameters:
      - $ref: "#/components/parameters/ActivityId"
    get:
      tags: [Activity]
      operationId: getPublicActivity
      summary: 获取公开活动详情
      description: |
        匿名可读取 PUBLISHED、CANCELLED 或 ENDED 活动。
        活动到达 endsAt 后，响应中的 status 和 registrationStatus 均为 ENDED。
        携带有效 Bearer 令牌时，同时返回本人参与状态和是否为活动发起者；匿名访问时本人状态为空。
        registrationGender 始终返回活动的报名性别限制；organizerPhone、organizerWechat 和
        organizerWechatQr 仅活动发起者或当前 ACTIVE 报名者可见，其他访问者返回 null。
        DRAFT、HIDDEN、无权访问和不存在统一按 404 处理，避免泄露非公开活动。
      x-implementation-status: IMPLEMENTED
      x-m3-extension-status: IMPLEMENTED
      security:
        - {}
        - bearerAuth: []
      responses:
        "200":
          description: 公开活动详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicActivityDetailResult"
        default:
          $ref: "#/components/responses/Error"

    delete:
      tags: [Activity]
      operationId: deleteActivityDraft
      summary: 删除本人或本企业活动草稿
      description: |
        仅资源所有者可操作，不接收请求体。只允许物理删除从未发布的 DRAFT。
        PUBLISHED、CANCELLED、ENDED 或 HIDDEN 返回 HTTP 409。
        资源不存在、重复删除或无管理权限统一返回 HTTP 404。
        删除活动、媒体关联和生命周期事件，但不直接删除文件对象。
        草稿删除后，原创建幂等键继续保留到原 24 小时窗口结束。
      x-implementation-status: IMPLEMENTED
      responses:
        "200":
          $ref: "#/components/responses/VoidSuccess"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
...
  /api/v1/users/me/activities:
    post:
      tags: [Activity]
      operationId: createMyActivity
      summary: 创建本人个人活动草稿
      description: |
        需要个人活动发布能力。创建固定返回 DRAFT，草稿至少提供非空标题。
        本接口只保存活动内容，不执行发布；发布使用统一 publication 命令。
        Idempotency-Key 在当前操作人、个人主体和键值范围内生效 24 小时。
        同键同请求重放返回原活动与 200；同键不同请求返回业务码 11601 和 HTTP 409。
        原幂等请求创建的草稿已删除时，同键同请求返回业务码 11605 和 HTTP 409。
      x-implementation-status: IMPLEMENTED
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonalActivityCreateRequest"
      responses:
        "200":
          description: 24 小时内同键同请求的幂等重放结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedActivityDetailResult"
        "201":
          description: 草稿已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedActivityDetailResult"
        "409":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/organizations/{organizationId}/activities:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
    post:
      tags: [Activity]
      operationId: createOrganizationActivity
      summary: 创建企业活动草稿
      description: |
        需要企业活动发布能力、启用企业和当前有效唯一 OWNER 关系。
        创建固定返回 DRAFT，草稿至少提供非空标题。
        本接口只保存活动内容，不执行发布；发布使用统一 publication 命令。
        Idempotency-Key 在当前操作人、企业主体和键值范围内生效 24 小时。
        同键不同请求返回业务码 11601 和 HTTP 409；原幂等请求创建的草稿已删除时，
        同键同请求返回业务码 11605 和 HTTP 409。
      x-implementation-status: IMPLEMENTED
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationActivityCreateRequest"
      responses:
        "200":
          description: 24 小时内同键同请求的幂等重放结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedActivityDetailResult"
        "201":
          description: 草稿已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedActivityDetailResult"
        "409":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/users/me/activities/{activityId}:
    parameters:
      - $ref: "#/components/parameters/ActivityId"
    put:
      tags: [Activity]
      operationId: updateMyActivity
      summary: 完整替换本人个人活动
      description: |
        仅允许本人活动。PUT 为完整替换并要求 version。
        DRAFT 编辑后保持 DRAFT；PUBLISHED 编辑后保持 PUBLISHED。
        发布必须调用统一 publication 命令。
        PUBLISHED 编辑继续满足全部发布字段约束。
        PUBLISHED 编辑时，原 endsAt 和替换后的 endsAt 都必须晚于服务端本次命令判定时刻；
        任一时间已到达时返回业务码 11602 和 HTTP 409。
        CANCELLED、ENDED 或 HIDDEN 不允许普通用户编辑；版本冲突返回业务码 11603 和 HTTP 409。
      x-implementation-status: IMPLEMENTED
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonalActivityUpdateRequest"
      responses:
        "200":
          description: 更新后的管理详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedActivityDetailResult"
        "409":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/organizations/{organizationId}/activities/{activityId}:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/ActivityId"
    put:
      tags: [Activity]
      operationId: updateOrganizationActivity
      summary: 完整替换本企业活动
      description: |
        仅允许启用企业的当前有效唯一 OWNER 操作本企业活动。PUT 为完整替换并要求 version。
        DRAFT 编辑后保持 DRAFT；PUBLISHED 编辑后保持 PUBLISHED。
        发布必须调用统一 publication 命令。
        PUBLISHED 编辑继续满足全部发布字段约束。
        PUBLISHED 编辑时，原 endsAt 和替换后的 endsAt 都必须晚于服务端本次命令判定时刻；
        任一时间已到达时返回业务码 11602 和 HTTP 409。
        CANCELLED、ENDED 或 HIDDEN 不允许普通用户编辑。
      x-implementation-status: IMPLEMENTED
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationActivityUpdateRequest"
      responses:
        "200":
          description: 更新后的管理详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedActivityDetailResult"
        "409":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/users/me/managed-activities:
    get:
      tags: [Activity]
      operationId: listMyManagedActivities
      summary: 游标分页查询当前账号可管理的活动
      description: |
        同时覆盖本人个人活动和当前账号可管理的企业活动。
        固定按 updatedAt DESC、activityId DESC 排序，cursor 是不透明字符串。
      x-implementation-status: IMPLEMENTED
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/ActivityLimit"
        - $ref: "#/components/parameters/ManagedActivityStatus"
        - $ref: "#/components/parameters/ActivityOwnerTypeFilter"
      responses:
        "200":
          description: 可管理活动分页
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedActivityPageResult"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/users/me/managed-activities/{activityId}:
    parameters:
      - $ref: "#/components/parameters/ActivityId"
    get:
      tags: [Activity]
      operationId: getMyManagedActivity
      summary: 获取当前账号可管理的活动详情
      description: |
        返回重新编辑所需的完整字段、真实发起主体摘要、当前报名人数、乐观锁版本和审计时间。
        不返回 operatorUserId、企业关系内部状态或敏感个人信息。
      x-implementation-status: IMPLEMENTED
      responses:
        "200":
          description: 活动管理详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedActivityDetailResult"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/activities/{activityId}/publication:
    parameters:
      - $ref: "#/components/parameters/ActivityId"
    put:
      tags: [Activity]
      operationId: publishActivity
      summary: 发布本人或本企业活动草稿
      description: |
        仅资源所有者可操作，不接收请求体。
        首次发布时，客户端先保存为 DRAFT：调用对应创建接口；编辑草稿后发布时，
        客户端先通过对应 PUT 保存最新内容，再调用本命令。
        当前 DRAFT 必须满足全部发布字段、时间顺序和图片规则，成功后转为 PUBLISHED。
        DRAFT 发布时重新校验当前操作人的资料完成度；资料不完整返回 11201 和 HTTP 409，活动保持 DRAFT。
        尚未到期（当前时间早于 endsAt）的 PUBLISHED 活动重复请求返回当前结果与 200，不重复写生命周期事件。
        已到达 endsAt 的活动按 ENDED 处理，返回 11602 和 HTTP 409。
        CANCELLED、ENDED 或 HIDDEN 返回 HTTP 409。
      x-implementation-status: IMPLEMENTED
      responses:
        "200":
          description: 发布后的活动管理详情，或已发布活动的幂等结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedActivityDetailResult"
        "409":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/activities/{activityId}/cancellation:
    parameters:
      - $ref: "#/components/parameters/ActivityId"
    put:
      tags: [Activity]
      operationId: cancelActivity
      summary: 取消本人或本企业已发布活动
      description: |
        仅资源所有者可操作，不接收请求体。允许 PUBLISHED 或 HIDDEN 转为 CANCELLED。
        取消成功时，在同一事务中将所有 ACTIVE 参与关系原子转为 TERMINATED，
        terminationReason 设为 ACTIVITY_CANCELLED，并将 participantCount 归零。
        已取消活动重复请求返回当前结果与 200，不重复写生命周期事件。
        DRAFT 不是可取消活动；DRAFT、ENDED，或已到达 endsAt 的 PUBLISHED/HIDDEN 返回业务码 11602 和 HTTP 409。
        有效结束后不再把保留的 ACTIVE 参与历史改写为 TERMINATED。
      x-implementation-status: IMPLEMENTED
      x-m3-extension-status: IMPLEMENTED
      responses:
        "200":
          description: 取消后的活动管理详情，或已取消活动的幂等结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedActivityDetailResult"
        "409":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    OrganizationId:
      name: organizationId
      in: path
      required: true
      description: 企业业务编号；JavaScript 客户端必须按字符串保存。
      schema:
        $ref: "#/components/schemas/Id"
    ActivityId:
      name: activityId
      in: path
      required: true
      description: 活动业务编号；JavaScript 客户端必须按字符串保存。
      schema:
        $ref: "#/components/schemas/Id"
...
    Cursor:
      name: cursor
      in: query
      required: false
      description: 后端生成的不透明分页游标，客户端不得解析或改写。
      schema:
        type: string
        minLength: 1
        maxLength: 512
...
    ActivityLimit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
...
    PublicActivityStatus:
      name: status
      in: query
      required: false
      description: 公开列表默认只查询已发布活动。
      schema:
        $ref: "#/components/schemas/PublicActivityStatus"
    ManagedActivityStatus:
      name: status
      in: query
      required: false
      description: 不传时查询当前账号可管理的全部状态。
      schema:
        $ref: "#/components/schemas/ActivityStatus"
    ActivityOwnerTypeFilter:
      name: ownerType
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/ActivityOwnerType"
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        创建命令的幂等键。结果在当前操作人、发布主体和键值范围内保留 24 小时。
        重放不延长窗口；活动或动态草稿删除也不缩短原 24 小时窗口。
        窗口到期后由请求事务先同步释放过期占用，再允许同键重新创建；后台清理延迟不得延长窗口。
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: "^[A-Za-z0-9._:-]+$"
  responses:
    VoidSuccess:
      description: 操作成功
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/VoidResult"
    Error:
      description: |
        统一错误响应。常见业务码：
        10000-10999 通用错误；11001-11005 登录会话；
        11101-11103 手机与设备；11201-11203 资料；
        11301-11302 协议；11401-11404 文件；11501-11503 账号数据；
        11601 幂等键冲突；11602 活动状态冲突；11603 乐观锁版本冲突；11604 活动容量冲突；
        11605 幂等请求对应的活动草稿已删除。
        11701 报名尚未开始；11702 报名已经截止；11703 活动名额已满；
        11704 活动发起者不能报名自己的活动；11705 活动已经开始；
        11706 活动已取消；11707 活动已结束；11708 当前参与记录不可由发起者移除；
        11709 已被活动发起者移除，不能再次报名；11710 不符合活动报名性别限制。
        11801 不能关注自己。
        11901 动态创建幂等键冲突；11902 动态状态冲突；11903 动态版本冲突；
        11904 幂等请求对应的动态草稿已删除；11905 动态内容未通过安全检查；
        11906 引用活动不可用。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResult"
...
  schemas:
    Id:
      type: string
      minLength: 1
      pattern: "^[1-9][0-9]*$"
      example: "1900000000000000201"
    Instant:
      type: string
      format: date-time
      example: "2026-07-29T08:00:00Z"
    ResultMeta:
      type: object
      required: [code, message, requestId, timestamp]
      properties:
        code:
          type: integer
          enum: [0]
        message:
          type: string
          example: 成功
        requestId:
          type: string
          example: 019fa7f8b1aa7253
        timestamp:
          $ref: "#/components/schemas/Instant"
    ErrorResult:
      type: object
      required: [code, message, data, requestId, timestamp]
      properties:
        code:
          type: integer
          minimum: 1
          example: 10001
        message:
          type: string
          example: 参数校验失败
        data:
          nullable: true
          description: 参数校验失败时可能为 FieldValidationError 数组。
        requestId:
          type: string
        timestamp:
          $ref: "#/components/schemas/Instant"
    VoidResult:
      allOf:
        - $ref: "#/components/schemas/ResultMeta"
        - type: object
          required: [data]
          properties:
            data:
              nullable: true
              example: null
...
    ActivityStatus:
      type: string
      enum: [DRAFT, PUBLISHED, CANCELLED, ENDED, HIDDEN]
    PublicActivityStatus:
      type: string
      enum: [PUBLISHED, CANCELLED, ENDED]
      default: PUBLISHED
    ActivityCategory:
      type: string
      enum:
        - HIKING
        - CAMPING
        - MOUNTAINEERING
        - RUNNING
        - CYCLING
        - FITNESS
        - BALL_SPORTS
        - WATER_SPORTS
        - TRAVEL
        - FOOD
        - MUSIC
        - MOVIE
        - READING
        - PHOTOGRAPHY
        - BOARD_GAMES
        - GAMING
        - PETS
        - PARENT_CHILD
        - VOLUNTEERING
        - OTHER
      description: 活动单选主分类；后续可以追加值，已有值原则上不改名或删除。
    ActivityRegionCode:
      type: string
      minLength: 6
      maxLength: 9
      pattern: "^[0-9]{6}([0-9]{3})?$"
      description: |
        编码来源为 AreaCity-JsSpider-StatsGov 2025.251231.260403 发布快照，使用项目
        规范化后的第三级节点原始 id。活动 regionCode 与个人资料 districtCode 使用同一份
        内置快照；后端仍须校验该编码对应实际区县级节点。
...
    ActivityLatitude:
      type: number
      format: double
      minimum: -90
      maximum: 90
      multipleOf: 0.0000001
      x-coordinate-reference-system: GCJ-02
      description: GCJ-02（国测局坐标系）的集合地点纬度，最多 7 位小数。
    ActivityLongitude:
      type: number
      format: double
      minimum: -180
      maximum: 180
      multipleOf: 0.0000001
      x-coordinate-reference-system: GCJ-02
      description: GCJ-02（国测局坐标系）的集合地点经度，最多 7 位小数。
...
    ActivityMap:
      type: object
      required: [items, truncated]
      properties:
        items:
          type: array
          maxItems: 200
          items:
            $ref: "#/components/schemas/ActivityMapItem"
        truncated:
          type: boolean
          description: true 表示当前地图视野内还有未返回的活动。
...
    PublicActivityPage:
      type: object
      required: [items, nextCursor, hasMore]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PublicActivitySummary"
        nextCursor:
          type: string
          nullable: true
        hasMore:
          type: boolean
...
    PublicActivityPageResult:
      allOf:
        - $ref: "#/components/schemas/ResultMeta"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/PublicActivityPage"
    ActivityMapResult:
      allOf:
        - $ref: "#/components/schemas/ResultMeta"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/ActivityMap"
    PublicActivityDetailResult:
      allOf:
        - $ref: "#/components/schemas/ResultMeta"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/PublicActivityDetail"
    ManagedActivityPageResult:
      allOf:
        - $ref: "#/components/schemas/ResultMeta"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/ManagedActivityPage"
    ManagedActivityDetailResult:
      allOf:
        - $ref: "#/components/schemas/ResultMeta"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/ManagedActivityDetail"
...
    PersonalActivityCreateRequest:
      type: object
      additionalProperties: false
      required: [title]
      description: |
        该接口固定创建 DRAFT，只保存发布待办内容，不执行发布；后续通过 publication 命令发布。
        草稿仅要求非空标题。发布命令会校验标题、分类、描述、封面、报名时间、活动时间、区县、
        详细地址和人数上限等公共发布字段完整，并校验以下时间顺序。
        时间必须满足 registrationStartsAt < registrationEndsAt <= startsAt < endsAt。
        coverFileId 不得出现在 mediaFileIds 中，数组省略时按空数组处理。
        所有图片必须由当前操作人上传、用途为 ACTIVITY 且处于可关联状态；关联后激活。
        字符串由服务端去除首尾空白，选填字符串处理后为空时保存为 null。
        latitude 与 longitude 必须同时为空或同时有值，非空时统一使用 GCJ-02；草稿允许同时为空，发布前必须同时有值。
        placeName 表示集合地点名称，草稿允许为空，发布时不得为空。
        coordinateSystem 在有坐标时可省略并默认 GCJ-02，显式值只接受 GCJ-02；无坐标时必须为空。
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 20
        categoryCode:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/ActivityCategory"
        description:
          type: string
          maxLength: 5000
          nullable: true
        coverFileId:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/Id"
        mediaFileIds:
          type: array
          maxItems: 5
          uniqueItems: true
          default: []
          items:
            $ref: "#/components/schemas/Id"
        registrationStartsAt:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/Instant"
        registrationEndsAt:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/Instant"
        startsAt:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/Instant"
        endsAt:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/Instant"
        regionCode:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/ActivityRegionCode"
        addressDetail:
          type: string
          minLength: 1
          maxLength: 512
          nullable: true
        placeName:
          $ref: "#/components/schemas/ActivityPlaceName"
        latitude:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/ActivityLatitude"
        longitude:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/ActivityLongitude"
        coordinateSystem:
          $ref: "#/components/schemas/ActivityCoordinateSystem"
        topics:
          $ref: "#/components/schemas/ActivityTopics"
        capacity:
          type: integer
          format: int32
          minimum: 1
          nullable: true
        signupDetails:
          type: string
          maxLength: 2000
          nullable: true
        organizerMessage:
          type: string
          maxLength: 1000
          nullable: true
        registrationGender:
          $ref: "#/components/schemas/ActivityRegistrationGender"
        organizerPhone:
          type: string
          maxLength: 32
          nullable: true
        organizerWechat:
          type: string
          maxLength: 64
          nullable: true
        organizerWechatQrFileId:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/Id"
        refundPolicy:
          type: string
          nullable: true
          description: 必须传 null；当前版本不启用退款政策。
...